Properly disable the cursor during device initialization
It seems the proprietary driver has a bug where it tries to set the
"cursor hide" bit in register 0xe868 using command 0x40/0xc8 (set cursor
position) instead of 0x40/0xc4 (set register value). I originally added
this code since the command was in the USB packet capture and I didn't
know what it did, but now that I do, and now that I know what it was
supposed to do, I'm changing this line to use the correct command.