Add experimental code for dissecting the video data
This is all a work-in-progress and is probably incorrect in many places.
The additional fields dissected by these changes should really just be
treated as aids to continue reverse engineering the protocol, and not as
a finished description of the protocol.

Normally I don't like to publish dissector code until I have some level
of confidence that my understanding of the protocol is correct. However,
in this case I'm publishing this WIP code before I'm confident that I
understand the protocol, and I'm doing this because I think it could
help others who are also currently reverse engineering this protocol.